+++ package.json
@@ -30,13 +30,17 @@
     "@expo/vector-icons": "^14.1.0",
     "@hookform/resolvers": "^5.2.1",
     "@openapitools/openapi-generator-cli": "^2.15.3",
+    "@react-native-async-storage/async-storage": "^2.2.0",
+    "@react-native-community/datetimepicker": "8.4.1",
     "@react-native-community/netinfo": "11.4.1",
     "@react-navigation/bottom-tabs": "^7.0.0",
     "@react-navigation/elements": "^2.2.0",
-    "@react-navigation/native": "^7.0.0",
+    "@react-navigation/native": "^7.1.17",
+    "@react-navigation/native-stack": "^7.3.25",
+    "@tanstack/react-query": "^5.85.0",
     "@testing-library/jest-native": "^5.4.3",
     "@testing-library/react-native": "^12.9.0",
-    "axios": "^1.7.8",
+    "axios": "^1.11.0",
     "dotenv": "^16.4.7",
     "expo": "~53.0.20",
     "expo-asset": "~11.1.7",
@@ -68,8 +72,7 @@
     "react-native-vector-icons": "^10.3.0",
     "react-native-web": "~0.20.0",
     "react-native-webview": "13.13.5",
-    "zod": "^4.0.17",
-    "@react-native-community/datetimepicker": "8.4.1"
+    "zod": "^4.0.17"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
