+++ src/app/_layout.tsx
@@ -1,21 +1,19 @@
 import { Stack } from "expo-router";
 import { StatusBar, StyleSheet } from "react-native";
 import { SafeAreaProvider } from "react-native-safe-area-context";
-
+import { QueryClientProvider, QueryClient } from "@tanstack/react-query";
 export default function RootLayout() {
+  const queryClient = new QueryClient();
   return (
     <SafeAreaProvider style={styles.container}>
-      <StatusBar barStyle="dark-content" />
-      <Stack screenOptions={{ headerShown: false, contentStyle: { backgroundColor: "white" } }}>
-        <Stack.Screen name="index" options={{ title: "LoginScreen" }} />
-      </Stack>
+      <QueryClientProvider client={queryClient}>
+        <StatusBar barStyle="dark-content" />
+        <Stack screenOptions={{ headerShown: false, contentStyle: { backgroundColor: "white" } }}>
+          <Stack.Screen name="index" options={{ title: "LoginScreen" }} />
+        </Stack>
+      </QueryClientProvider>
     </SafeAreaProvider>
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#FFFFFF"
-  }
-});
+const styles = StyleSheet.create({ container: { flex: 1, backgroundColor: "#FFFFFF" } });
