+++ src/shared/apis/api/sign-up-api.ts
@@ -1,0 +1,6 @@
+import api from "../base";
+import { SignUpRequest } from "../model";
+
+export const postSignUp = (payload: SignUpRequest) => {
+  return api.post("/members/save");
+};
