+++ src/shared/apis/index.ts
@@ -13,5 +13,4 @@
  */
 
 export * from "./api";
-export * from "./configuration";
 export * from "./model";
