+++ src/shared/apis/model/index.ts
@@ -1,8 +1,1 @@
-export * from "./address";
-export * from "./api-response";
-export * from "./category";
-export * from "./customer";
-export * from "./order";
-export * from "./pet";
-export * from "./tag";
 export * from "./user";
