--- src/shared/services/apis/README.md
@@ -1,1 +1,0 @@
-# shared > services > apis
